Extract sdk from ai gateway adapter
